refactor: 도메인 클래스를 domain 패키지 내부로 이동
